// frontend/app/src/app/model/user.model.ts
export class User {
    username: string;
    password: string;
    name: string;
    surname: string;
    index: string;
    type_st: string;
    address: string;
    phone: string;
    web: string;
    title: string;
    cabinet: string;
    changePass: number;
    status: number;
    type: string;
    predmeti: Array<Object>;
}